+++ lbrynet/lbryfilemanager/EncryptedFileManager.py
@@ -9,6 +9,7 @@ from twisted.enterprise import adbapi
 from twisted.internet import defer, task, reactor
 from twisted.python.failure import Failure
 
+from lbrynet.reflector.reupload import reflect_stream
 from lbrynet.core.PaymentRateManager import NegotiatedPaymentRateManager
 from lbrynet.lbryfilemanager.EncryptedFileDownloader import ManagedEncryptedFileDownloader
 from lbrynet.lbryfilemanager.EncryptedFileDownloader import ManagedEncryptedFileDownloaderFactory
@@ -19,6 +20,16 @@ from lbrynet.core.sqlite_helpers import rerun_if_locked
 
 
 log = logging.getLogger(__name__)
+
+
+def safe_start_looping_call(looping_call, seconds=3600):
+    if not looping_call.running:
+        looping_call.start(seconds)
+
+
+def safe_stop_looping_call(looping_call):
+    if looping_call.running:
+        looping_call.stop()
 
 
 class EncryptedFileManager(object):
@@ -32,11 +43,13 @@ class EncryptedFileManager(object):
         self.stream_info_manager = stream_info_manager
         self.sd_identifier = sd_identifier
         self.lbry_files = []
+        self.lbry_files_setup_deferred = None
         self.sql_db = None
         if download_directory:
             self.download_directory = download_directory
         else:
             self.download_directory = os.getcwd()
+        self.lbry_file_reflector = task.LoopingCall(self.reflect_lbry_files)
         log.debug("Download directory for EncryptedFileManager: %s", str(self.download_directory))
 
     @defer.inlineCallbacks
@@ -44,6 +57,7 @@ class EncryptedFileManager(object):
         yield self._open_db()
         yield self._add_to_sd_identifier()
         yield self._start_lbry_files()
+        safe_start_looping_call(self.lbry_file_reflector)
 
     def get_lbry_file_status(self, lbry_file):
         return self._get_lbry_file_status(lbry_file.rowid)
@@ -68,6 +82,9 @@ class EncryptedFileManager(object):
 
         dl.addCallback(filter_failures)
         return dl
+
+    def save_sd_blob_hash_to_stream(self, stream_hash, sd_hash):
+        return self.stream_info_manager.save_sd_blob_hash_to_stream(stream_hash, sd_hash)
 
     def _add_to_sd_identifier(self):
         downloader_factory = ManagedEncryptedFileDownloaderFactory(self)
@@ -97,64 +114,77 @@ class EncryptedFileManager(object):
         yield defer.DeferredList(list(_iter_streams(stream_hashes)))
 
     @defer.inlineCallbacks
+    def _restore_lbry_file(self, lbry_file):
+        try:
+            yield lbry_file.restore()
+        except Exception as err:
+            log.error("Failed to start stream: %s, error: %s", lbry_file.stream_hash, err)
+            self.lbry_files.remove(lbry_file)
+            # TODO: delete stream without claim instead of just removing from manager?
+
+    @defer.inlineCallbacks
     def _start_lbry_files(self):
-        def set_options_and_restore(rowid, stream_hash, options):
-            b_prm = self.session.base_payment_rate_manager
-            payment_rate_manager = NegotiatedPaymentRateManager(b_prm, self.session.blob_tracker)
-
-            d = self.start_lbry_file(rowid, stream_hash, payment_rate_manager,
-                                     blob_data_rate=options)
-            d.addCallback(lambda downloader: downloader.restore())
-            return d
-
-        def log_error(err, rowid, stream_hash, options):
-            log.error("An error occurred while starting a lbry file: %s", err.getErrorMessage())
-            log.debug(rowid)
-            log.debug(stream_hash)
-            log.debug(options)
-
-        def start_lbry_files(lbry_files_and_options):
-            for rowid, stream_hash, options in lbry_files_and_options:
-                d = set_options_and_restore(rowid, stream_hash, options)
-                d.addErrback(lambda err: log_error(err, rowid, stream_hash, options))
-            log.info("Started %i lbry files", len(self.lbry_files))
-            return True
-
+        b_prm = self.session.base_payment_rate_manager
+        payment_rate_manager = NegotiatedPaymentRateManager(b_prm, self.session.blob_tracker)
         yield self._check_stream_info_manager()
-        files_and_options = yield self._get_all_lbry_files()
-        yield start_lbry_files(files_and_options)
+        lbry_files_and_options = yield self._get_all_lbry_files()
+        dl = []
+        for rowid, stream_hash, options in lbry_files_and_options:
+            lbry_file = yield self.start_lbry_file(rowid, stream_hash, payment_rate_manager,
+                                                   blob_data_rate=options)
+            dl.append(self._restore_lbry_file(lbry_file))
+            log.debug("Started %s", lbry_file)
+        self.lbry_files_setup_deferred = defer.DeferredList(dl)
+        log.info("Started %i lbry files", len(self.lbry_files))
+        defer.returnValue(True)
 
+    @defer.inlineCallbacks
     def start_lbry_file(self, rowid, stream_hash,
                         payment_rate_manager, blob_data_rate=None, upload_allowed=True,
                         download_directory=None, file_name=None):
         if not download_directory:
             download_directory = self.download_directory
         payment_rate_manager.min_blob_data_payment_rate = blob_data_rate
-        lbry_file_downloader = ManagedEncryptedFileDownloader(rowid, stream_hash,
-                                                         self.session.peer_finder,
-                                                         self.session.rate_limiter,
-                                                         self.session.blob_manager,
-                                                         self.stream_info_manager, self,
-                                                         payment_rate_manager, self.session.wallet,
-                                                         download_directory,
-                                                         upload_allowed,
-                                                         file_name=file_name)
-        self.lbry_files.append(lbry_file_downloader)
-        d = lbry_file_downloader.set_stream_info()
-        d.addCallback(lambda _: lbry_file_downloader)
-        return d
+        lbry_file = ManagedEncryptedFileDownloader(rowid, stream_hash, self.session.peer_finder,
+                                                   self.session.rate_limiter,
+                                                   self.session.blob_manager,
+                                                   self.stream_info_manager,
+                                                   self, payment_rate_manager, self.session.wallet,
+                                                   download_directory, upload_allowed,
+                                                   file_name=file_name)
+        yield lbry_file.set_stream_info()
+        self.lbry_files.append(lbry_file)
+        defer.returnValue(lbry_file)
 
-    def add_lbry_file(self, stream_hash, payment_rate_manager,
-                      blob_data_rate=None,
-                      upload_allowed=True,
-                      download_directory=None,
-                      file_name=None):
-        d = self._save_lbry_file(stream_hash, blob_data_rate)
-        d.addCallback(
-            lambda rowid: self.start_lbry_file(
-                rowid, stream_hash, payment_rate_manager,
-                blob_data_rate, upload_allowed, download_directory, file_name))
-        return d
+    @defer.inlineCallbacks
+    def _stop_lbry_file(self, lbry_file):
+        def wait_for_finished(lbry_file, count=2):
+            if count or lbry_file.saving_status is not False:
+                return task.deferLater(reactor, 1, self._stop_lbry_file, lbry_file, count=count - 1)
+        try:
+            yield lbry_file.stop(change_status=False)
+            self.lbry_files.remove(lbry_file)
+        except CurrentlyStoppingError:
+            yield wait_for_finished(lbry_file)
+        except AlreadyStoppedError:
+            pass
+        finally:
+            defer.returnValue(None)
+
+    def _stop_lbry_files(self):
+        log.info("Stopping %i lbry files", len(self.lbry_files))
+        lbry_files = self.lbry_files
+        for lbry_file in lbry_files:
+            yield self._stop_lbry_file(lbry_file)
+
+    @defer.inlineCallbacks
+    def add_lbry_file(self, stream_hash, payment_rate_manager, blob_data_rate=None,
+                      upload_allowed=True, download_directory=None, file_name=None):
+        rowid = yield self._save_lbry_file(stream_hash, blob_data_rate)
+        lbry_file = yield self.start_lbry_file(rowid, stream_hash, payment_rate_manager,
+                                               blob_data_rate, upload_allowed, download_directory,
+                                               file_name)
+        defer.returnValue(lbry_file)
 
     def delete_lbry_file(self, lbry_file):
         for l in self.lbry_files:
@@ -192,31 +222,22 @@ class EncryptedFileManager(object):
         else:
             return defer.fail(Failure(ValueError("Could not find that LBRY file")))
 
-    def stop(self):
-        log.info('Stopping %s', self)
-        ds = []
-
-        def wait_for_finished(lbry_file, count=2):
-            if count <= 0 or lbry_file.saving_status is False:
-                return True
-            else:
-                return task.deferLater(reactor, 1, wait_for_finished, lbry_file, count=count - 1)
-
-        def ignore_stopped(err, lbry_file):
-            err.trap(AlreadyStoppedError, CurrentlyStoppingError)
-            return wait_for_finished(lbry_file)
-
+    def _reflect_lbry_files(self):
         for lbry_file in self.lbry_files:
-            d = lbry_file.stop(change_status=False)
-            d.addErrback(ignore_stopped, lbry_file)
-            ds.append(d)
-        dl = defer.DeferredList(ds)
+            yield reflect_stream(lbry_file)
 
-        def close_db():
-            self.db = None
+    @defer.inlineCallbacks
+    def reflect_lbry_files(self):
+        yield defer.DeferredList(list(self._reflect_lbry_files()))
 
-        dl.addCallback(lambda _: close_db())
-        return dl
+    @defer.inlineCallbacks
+    def stop(self):
+        safe_stop_looping_call(self.lbry_file_reflector)
+        yield defer.DeferredList(list(self._stop_lbry_files()))
+        yield self.sql_db.close()
+        self.sql_db = None
+        log.info("Stopped %s", self)
+        defer.returnValue(True)
 
     def get_count_for_stream_hash(self, stream_hash):
         return self._get_count_for_stream_hash(stream_hash)